+++ src/App.tsx
@@ -1,13 +1,17 @@
 import React from 'react'
 import './App.css'
 import { Button } from './components/button/Button'
-import LoginPage from './pages/LoginPage/LoginPage'
+import FormRegistration from './components/Forms/FormRegistration'
+import FormResetPassword from './components/Forms/FormResetPassword'
+import FormLogin from './components/Forms/FormLogin'
 
 function App() {
   return (
     <>
     {/* <Button /> */}
-    <LoginPage />
+    <FormRegistration />
+    {/* <FormResetPassword /> */}
+    {/* <FormLogin /> */}
     </>
   )
 }
